Resolve comments, and override next-auth default user model that's put in database when user signs up with SSO
